% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpcaResult-class.R
\name{ContrastCondName,tpcaResult-method}
\alias{ContrastCondName,tpcaResult-method}
\alias{ContrastCondName}
\title{Extract ContrastCondName}
\usage{
\S4method{ContrastCondName}{tpcaResult}(object)
}
\arguments{
\item{object}{and object of class tpcaResult}
}
\value{
a character string describing the contrast 
condition
}
\description{
Extract ContrastCondName
}
\examples{
tpcaObj <- new("tpcaResult")
ContrastCondName(tpcaObj)
}
