% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpcaResult-class.R
\name{CtrlCondName,tpcaResult-method}
\alias{CtrlCondName,tpcaResult-method}
\alias{CtrlCondName}
\title{Extract CtrlCondName}
\usage{
\S4method{CtrlCondName}{tpcaResult}(object)
}
\arguments{
\item{object}{and object of class tpcaResult}
}
\value{
a character string describing the control 
condition
}
\description{
Extract CtrlCondName
}
\examples{
tpcaObj <- new("tpcaResult")
CtrlCondName(tpcaObj)
}
