% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpcaResult-class.R
\name{DistMethod,tpcaResult-method}
\alias{DistMethod,tpcaResult-method}
\alias{DistMethod}
\title{Extract DistMethod}
\usage{
\S4method{DistMethod}{tpcaResult}(object)
}
\arguments{
\item{object}{and object of class tpcaResult}
}
\value{
a character string of the dist method
}
\description{
Extract DistMethod
}
\examples{
tpcaObj <- new("tpcaResult")
DistMethod(tpcaObj)
}
