% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ori_et_al_complex_ppis}
\alias{ori_et_al_complex_ppis}
\title{Data frame of eukaryotic protein-protein
interactions inferred from annotated protein 
complexes by Ori et al.}
\format{data frame with columns complex_name, x, y, 
pair (unique pair id)}
\usage{
data("ori_et_al_complex_ppis")
}
\description{
data frame assigning proteins to (in)directly 
interacting proteins within protein complexes
}
\examples{
data("ori_et_al_complex_ppis")
}
\references{
Ori et al. (2016), Genome Biology, 17, 47
}
\keyword{datasets}
