% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ori_et_al_complexes_df}
\alias{ori_et_al_complexes_df}
\title{Data frame of annotated protein complexes by Ori 
et al.}
\format{
data frame with columns ensembl_id, protein
and id (complex identifier)
}
\usage{
data("ori_et_al_complexes_df")
}
\description{
data frame assigning proteins to 
annotated protein complexes
}
\examples{
data("ori_et_al_complexes_df")
}
\references{
Ori et al. (2016), Genome Biology, 17, 47
}
\keyword{datasets}
