% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpcaResult-class.R
\name{PPiRocTableAnno,tpcaResult-method}
\alias{PPiRocTableAnno,tpcaResult-method}
\alias{PPiRocTableAnno}
\title{Extract PPiRocTableAnno}
\usage{
\S4method{PPiRocTableAnno}{tpcaResult}(object)
}
\arguments{
\item{object}{an object of class tpcaResult}
}
\value{
a data frame containing annotation 
information for PPiRocTable
}
\description{
Extract PPiRocTableAnno
}
\examples{
m1 <- matrix(1:12, ncol = 4)
m2 <- matrix(2:13, ncol = 4)
m3 <- matrix(c(2:10, 1:7), ncol = 4)

rownames(m1) <- 1:3
rownames(m2) <- 2:4
rownames(m3) <- 2:5

mat_list <- list(
    m1, m2, m3
)
tpcaObj <- new("tpcaResult", ObjList = mat_list)
PPiRocTableAnno(tpcaObj)
}
