% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{runTPCA}
\alias{runTPCA}
\title{Run the TPCA analysis}
\usage{
runTPCA(objList, complexAnno = NULL, ppiAnno = NULL,
  rownameCol = NULL, summaryFUN = median, distMethod = "euclidean",
  stringScore = 700)
}
\arguments{
\item{objList}{inout list of objects, e.g.
ExpressionSets retrieved after TPP data import
or matrices or data frames}

\item{complexAnno}{data frame annotating
known protein complexes of interest to test}

\item{ppiAnno}{data frame annotation known
protein-protein interactions (PPI) to test}

\item{rownameCol}{in case the input objects are tibbles
this parameter takes in the name (character) of the column 
specifying protein names or ids}

\item{distMethod}{method to use within dist function,
default is 'euclidean'}

\item{stringScore}{score cutoff to filter
PPIs retrieved from StringDb}

\item{summaryFun}{function to use to summarize measurements
across replicates, default is median}
}
\value{
an object of class tpcaResult
with the following slots:
1) ObjList: containing the supplied list of
objects

 @export
}
\description{
Run the TPCA analysis
}
\examples{
 
}
