% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpcaResult-class.R
\name{SetComplexBackgroundDistributionList,tpcaResult-method}
\alias{SetComplexBackgroundDistributionList,tpcaResult-method}
\alias{SetComplexBackgroundDistributionList}
\title{Set ComplexBackgroundDistributionList}
\usage{
\S4method{SetComplexBackgroundDistributionList}{tpcaResult}(object, lt)
}
\arguments{
\item{object}{an object of class tpcaResult}

\item{lt}{a list of data frames containing distances
of random complexes with different number of members}
}
\value{
an object of class tpcaResult
}
\description{
Set ComplexBackgroundDistributionList
}
\examples{
m1 <- matrix(1:12, ncol = 4)
m2 <- matrix(2:13, ncol = 4)
m3 <- matrix(c(2:10, 1:7), ncol = 4)

rownames(m1) <- 1:3
rownames(m2) <- 2:4
rownames(m3) <- 2:5

mat_list <- list(
    m1, m2, m3
)
tpcaObj <- new("tpcaResult", ObjList = mat_list)
SetComplexBackgroundDistributionList(tpcaObj, 
  list('3' = data.frame(pair = "A:B")))
}
