% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpcaResult-class.R
\name{SetContrastCondName,tpcaResult-method}
\alias{SetContrastCondName,tpcaResult-method}
\alias{SetContrastCondName}
\title{Set ContrastCondName}
\usage{
\S4method{SetContrastCondName}{tpcaResult}(object, name)
}
\arguments{
\item{object}{an object of class tpcaResult}

\item{name}{a character string describing the contrast 
condition}
}
\value{
an object of class tpcaResult
}
\description{
Set ContrastCondName
}
\examples{
tpcaObj <- new("tpcaResult")
SetContrastCondName(tpcaObj, "DMSO")
}
