% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpcaResult-class.R
\name{SetContrastDistMat,tpcaResult-method}
\alias{SetContrastDistMat,tpcaResult-method}
\alias{SetContrastDistMat}
\title{Set ContrastDistMat}
\usage{
\S4method{SetContrastDistMat}{tpcaResult}(object, mat)
}
\arguments{
\item{object}{and object of class tpcaResult}

\item{mat}{matrix containg contrast distance 
matrix to set}
}
\value{
a data frame containing the results
from a tpca analysis
}
\description{
Set ContrastDistMat
}
\examples{
m1 <- matrix(1:12, ncol = 4)
m2 <- matrix(2:13, ncol = 4)
m3 <- matrix(c(2:10, 1:7), ncol = 4)

rownames(m1) <- 1:3
rownames(m2) <- 2:4
rownames(m3) <- 2:5

mat_list <- list(
    m1, m2, m3
)
tpcaObj <- new("tpcaResult", ObjList = mat_list)
SetContrastDistMat(tpcaObj, matrix(c(0, 1, 0, 1), ncol = 2))
}
