% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpcaResult-class.R
\name{SetCtrlCondName,tpcaResult-method}
\alias{SetCtrlCondName,tpcaResult-method}
\alias{SetCtrlCondName}
\title{Set CtrlCondName}
\usage{
\S4method{SetCtrlCondName}{tpcaResult}(object, name)
}
\arguments{
\item{object}{an object of class tpcaResult}

\item{name}{a character string describing the control 
condition}
}
\value{
an object of class tpcaResult
}
\description{
Set CtrlCondName
}
\examples{
tpcaObj <- new("tpcaResult")
SetCtrlCondName(tpcaObj, "DMSO")
}
