% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpcaResult-class.R
\name{SetDistMethod,tpcaResult-method}
\alias{SetDistMethod,tpcaResult-method}
\alias{SetDistMethod}
\title{Set distMethod}
\usage{
\S4method{SetDistMethod}{tpcaResult}(object, method)
}
\arguments{
\item{object}{an object of class tpcaResult}

\item{method}{character string of dist method}
}
\value{
an object of class tpcaResult
}
\description{
Set distMethod
}
\examples{
tpcaObj <- new("tpcaResult")
SetDistMethod(tpcaObj, "euclidean")
}
