% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpcaResult-class.R
\name{SetPPiRocTable,tpcaResult-method}
\alias{SetPPiRocTable,tpcaResult-method}
\alias{SetPPiRocTable}
\title{Set PPiRocTable}
\usage{
\S4method{SetPPiRocTable}{tpcaResult}(object, df)
}
\arguments{
\item{object}{an object of class tpcaResult}

\item{df}{data.frame containg PPiRocTable to set}
}
\value{
an object of class tpcaResult
}
\description{
Set PPiRocTable
}
\examples{
m1 <- matrix(1:12, ncol = 4)
m2 <- matrix(2:13, ncol = 4)
m3 <- matrix(c(2:10, 1:7), ncol = 4)

rownames(m1) <- 1:3
rownames(m2) <- 2:4
rownames(m3) <- 2:5

mat_list <- list(
    m1, m2, m3
)
tpcaObj <- new("tpcaResult", ObjList = mat_list)
SetPPiRocTable(tpcaObj, data.frame(FPR = 1, TPR = 1))
}
