% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpcaResult-class.R
\name{SetSummaryMethod,tpcaResult-method}
\alias{SetSummaryMethod,tpcaResult-method}
\alias{SetSummaryMethod}
\title{Set summaryMethod}
\usage{
\S4method{SetSummaryMethod}{tpcaResult}(object, method)
}
\arguments{
\item{object}{an object of class tpcaResult}

\item{method}{character string of summary method}
}
\value{
an object of class tpcaResult
}
\description{
Set summaryMethod
}
\examples{
tpcaObj <- new("tpcaResult")
SetSummaryMethod(tpcaObj, "median")
}
