% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpcaResult-class.R
\name{SetTpcaResultTable,tpcaResult-method}
\alias{SetTpcaResultTable,tpcaResult-method}
\alias{SetTpcaResultTable}
\title{Set TpcaResultTable}
\usage{
\S4method{SetTpcaResultTable}{tpcaResult}(object, df)
}
\arguments{
\item{object}{an object of class tpcaResult}

\item{df}{a data frame containing the results
from a tpca analysis}
}
\value{
an object of class tpcaResult
}
\description{
Set TpcaResultTable
}
\examples{
m1 <- matrix(1:12, ncol = 4)
m2 <- matrix(2:13, ncol = 4)
m3 <- matrix(c(2:10, 1:7), ncol = 4)

rownames(m1) <- 1:3
rownames(m2) <- 2:4
rownames(m3) <- 2:5

mat_list <- list(
    m1, m2, m3
)
tpcaObj <- new("tpcaResult", ObjList = mat_list)
SetTpcaResultTable(tpcaObj, data.frame(pair = "A:B"))
}
