% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{string_ppi_df}
\alias{string_ppi_df}
\title{Data frame of annotated human protein-protein
interactions retrieved from stringDB with a combined
interaciton score euqal or higher than 700}
\format{
data frame with columns x, y (gene symbol of 
interactors), combined_score, pair (unique pair id)
}
\usage{
data("string_ppi_df")
}
\description{
data frame assigning proteins to 
interacting proteins
}
\examples{
data("string_ppi_df")
}
\references{
Jensen et al. (2009), Nucleic Acids Research, 
37, D412–D416
}
\keyword{datasets}
