% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpcaResult-class.R
\name{SummaryMethod,tpcaResult-method}
\alias{SummaryMethod,tpcaResult-method}
\alias{SummaryMethod}
\title{Extract SummaryMethod}
\usage{
\S4method{SummaryMethod}{tpcaResult}(object)
}
\arguments{
\item{object}{and object of class tpcaResult}
}
\value{
a character string of the summary method
}
\description{
Extract SummaryMethod
}
\examples{
tpcaObj <- new("tpcaResult")
SummaryMethod(tpcaObj)
}
