% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpcaResult-class.R
\name{tpcaResultTable,tpcaResult-method}
\alias{tpcaResultTable,tpcaResult-method}
\alias{tpcaResultTable}
\title{Extract tpcaResultTable}
\usage{
\S4method{tpcaResultTable}{tpcaResult}(object)
}
\arguments{
\item{object}{and object of class tpcaResult}
}
\value{
a data frame containing the results
from a tpca analysis
}
\description{
Extract tpcaResultTable
}
\examples{
m1 <- matrix(1:12, ncol = 4)
m2 <- matrix(2:13, ncol = 4)
m3 <- matrix(c(2:10, 1:7), ncol = 4)

rownames(m1) <- 1:3
rownames(m2) <- 2:4
rownames(m3) <- 2:5

mat_list <- list(
    m1, m2, m3
)
tpcaObj <- new("tpcaResult", ObjList = mat_list)
tpcaResultTable(tpcaObj)
}
