% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{runTPCA}
\alias{runTPCA}
\title{Run the TPCA analysis}
\usage{
runTPCA(objList, complexAnno = NULL, ppiAnno = NULL,
  rownameCol = NULL, summaryFUN = median, distMethod = "euclidean",
  stringScore = 700)
}
\arguments{
\item{objList}{inout list of objects, e.g.
ExpressionSets retrieved after TPP data import
or matrices or data frames}

\item{complexAnno}{data frame annotating
known protein complexes of interest to test}

\item{ppiAnno}{data frame annotation known
protein-protein interactions (PPI) to test}

\item{rownameCol}{in case the input objects are tibbles
this parameter takes in the name (character) of the column 
specifying protein names or ids}

\item{summaryFUN}{function to use to summarize measurements
across replicates, default is median}

\item{distMethod}{method to use within dist function,
default is 'euclidean'}

\item{stringScore}{score cutoff to filter
PPIs retrieved from StringDb}
}
\value{
an object of class tpcaResult
with the following slots:
1) ObjList: containing the supplied list of
objects
}
\description{
Run the TPCA analysis
}
\examples{
 
m1 <- matrix(1:12, ncol = 4)
m2 <- matrix(2:13, ncol = 4)
m3 <- matrix(c(2:10, 1:7), ncol = 4)

rownames(m1) <- 1:3
rownames(m2) <- 2:4
rownames(m3) <- 2:5

colnames(m1) <- paste0("X", 1:4)
colnames(m2) <- paste0("X", 1:4)
colnames(m3) <- paste0("X", 1:4)

mat_list <- list(
    m1, m2, m3
)

ppi_anno <- tibble(
    gene_name1 = "2",
    gene_name2 = "3",
    combined_score = 700,
    pair = "2:3")

runTPCA(
    objList = mat_list,
    complexAnno = NULL,
    ppiAnno = ppi_anno
)
}
