% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpcaResult-class.R
\docType{class}
\name{tpcaResult-class}
\alias{tpcaResult-class}
\alias{tpcaResult}
\title{S4 TPCA Result Class}
\value{
an object of class tpcaResult
with the following slots:
1) ObjList: containing the supplied list of
objects
}
\description{
S4 TPCA Result Class
}
\section{Slots}{

\describe{
\item{\code{ObjList}}{list.}

\item{\code{ContrastList}}{list.}

\item{\code{DistMat}}{matrix.}

\item{\code{ContrastDistMat}}{matrix.}

\item{\code{ComplexAnnotation}}{data.frame.}

\item{\code{PPiAnnotation}}{data.frame.}

\item{\code{PPiRocTable}}{data.frame.}

\item{\code{summaryFUN}}{function.}

\item{\code{distMethod}}{character.}
}}

\examples{
m1 <- matrix(1:12, ncol = 4)
m2 <- matrix(2:13, ncol = 4)
m3 <- matrix(c(2:10, 1:7), ncol = 4)

rownames(m1) <- 1:3
rownames(m2) <- 2:4
rownames(m3) <- 2:5

mat_list <- list(
    m1, m2, m3
)
tpcaObj <- new("tpcaResult", ObjList = mat_list)
}
